+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ui-dev
 
+## 1.0.19
+
+### Patch Changes
+
+- - update alert, colorarea documentation
+  - add field, radio documentation
+
 ## 1.0.18
 
 ### Patch Changes