+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @elf-framework/design-system
 
+## 2.1.24
+
+### Patch Changes
+
+- - modify Alert style
+  - support Field, TextField component
+  - support RadioGroup component
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.1.24
+
 ## 2.1.22
 
 ### Patch Changes