+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @elf-framework/design-tokens
 
+## 2.1.24
+
+### Patch Changes
+
+- - modify Alert style
+  - support Field, TextField component
+  - support RadioGroup component
+
 ## 2.1.22
 
 ### Patch Changes